messaging_app/chats/permissions.py
from rest_framework.permissions import BasePermission

class IsParticipantOfConversation(BasePermission):
    """
    Allows access only to participants of the conversation.
    """

    def has_object_permission(self, request, view, obj):
        # Check if the user is in the conversation participants
        return request.user in obj.participants.all()
